ai/requirements.txt
flask[async]
requests
python-dotenv
openai
aiohttp